# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Yral AI Chat API
  description: AI Chat API for Yral with multimodal support
  version: 1.0.0
paths:
...
  /api/v1/chat/conversations:
    post:
      tags:
      - Chat
      summary: Create a new conversation
      description: "Create a new conversation with an AI influencer.\n    \n    If\
        \ a conversation already exists between the user and influencer, \n    returns\
        \ the existing conversation instead of creating a new one.\n    \n    **Authentication\
        \ required**: JWT token in Authorization header"
      operationId: createConversation
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
      responses:
        '201':
          description: Conversation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
              example:
                id: 550e8400-e29b-41d4-a716-446655440000
                user_id: user123
                influencer:
                  id: 660e8400-e29b-41d4-a716-446655440000
                  name: tech_guru
                  display_name: Tech Guru AI
                  avatar_url: https://example.com/avatar.jpg
                created_at: '2024-01-15T10:30:00Z'
                updated_at: '2024-01-15T10:30:00Z'
                message_count: 0
        '400':
          description: Bad request - Invalid influencer_id format
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '404':
          description: Influencer not found
        '422':
          description: Validation error - Request body validation failed
        '429':
          description: Rate limit exceeded
        '500':
          description: Internal server error
    get:
      tags:
      - Chat
      summary: List user conversations
      description: Retrieve paginated list of user's conversations, optionally filtered
        by influencer
      operationId: listConversations
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Number of conversations to return
          default: 20
          title: Limit
        description: Number of conversations to return
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Number of conversations to skip
          default: 0
          title: Offset
        description: Number of conversations to skip
      - name: influencer_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Filter by specific influencer ID
          title: Influencer Id
        description: Filter by specific influencer ID
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: List of conversations retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListConversationsResponse'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '422':
          description: Validation error - Invalid query parameters
        '429':
          description: Rate limit exceeded
        '500':
          description: Internal server error
  /api/v1/chat/conversations/{conversation_id}/messages:
    get:
      tags:
      - Chat
      summary: Get conversation messages
      description: Retrieve paginated message history for a conversation
      operationId: listMessages
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Conversation Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Number of messages to return
          default: 50
          title: Limit
        description: Number of messages to return
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Number of messages to skip
          default: 0
          title: Offset
        description: Number of messages to skip
      - name: order
        in: query
        required: false
        schema:
          type: string
          pattern: ^(asc|desc)$
          description: 'Sort order: ''asc'' or ''desc'''
          default: desc
          title: Order
        description: 'Sort order: ''asc'' or ''desc'''
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMessagesResponse'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - Not authorized to access this conversation
        '404':
          description: Conversation not found
        '422':
          description: Validation error - Invalid parameters
        '429':
          description: Rate limit exceeded
        '500':
          description: Internal server error
    post:
      tags:
      - Chat
      summary: Send a message
      description: "Send a message to an AI influencer and receive a response.\n \
        \   \n    Supports multiple message types:\n    - **TEXT**: Plain text messages\n\
        \    - **IMAGE**: Image-only messages  \n    - **MULTIMODAL**: Text with images\n\
        \    - **AUDIO**: Voice/audio messages\n    \n    The AI response is returned\
        \ immediately. Background tasks handle logging and cache updates."
      operationId: sendMessage
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Conversation Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: Message sent and AI response received successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageResponse'
        '400':
          description: Bad request - Invalid message format or content
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - Not authorized to access this conversation
        '404':
          description: Conversation not found
        '422':
          description: Validation error - Invalid message data
        '429':
          description: Rate limit exceeded
        '500':
          description: Internal server error
        '503':
          description: Service unavailable - AI service temporarily unavailable
  /api/v1/chat/conversations/{conversation_id}:
    delete:
      tags:
      - Chat
      summary: Delete a conversation
      description: Permanently delete a conversation and all its associated messages
      operationId: deleteConversation
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Conversation Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Conversation deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteConversationResponse'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - Not authorized to delete this conversation
        '404':
          description: Conversation not found
        '429':
          description: Rate limit exceeded
        '500':
          description: Internal server error
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ConversationResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          title: User Id
        influencer:
          $ref: '#/components/schemas/InfluencerBasicInfo'
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        message_count:
          type: integer
          title: Message Count
          default: 0
        last_message:
          anyOf:
          - $ref: '#/components/schemas/LastMessageInfo'
          - type: 'null'
          title: Last Message
        recent_messages:
          anyOf:
          - type: array
            items:
              $ref: '#/components/schemas/MessageResponse'
          - type: 'null'
          title: Recent Messages
      type: object
      required:
      - id
      - user_id
      - influencer
      - created_at
      - updated_at
      title: ConversationResponse
      description: Conversation response model
    CreateConversationRequest:
      properties:
        influencer_id:
          type: string
          format: uuid
          title: Influencer Id
          description: ID of the AI influencer
          examples:
          - 550e8400-e29b-41d4-a716-446655440000
      type: object
      required:
      - influencer_id
      title: CreateConversationRequest
      description: Request to create a new conversation
      examples:
      - influencer_id: 550e8400-e29b-41d4-a716-446655440000
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ListConversationsResponse:
      properties:
        conversations:
          items:
            $ref: '#/components/schemas/ConversationResponse'
          type: array
          title: Conversations
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - conversations
      - total
      - limit
      - offset
      title: ListConversationsResponse
      description: Response for listing conversations
...
    ListMessagesResponse:
      properties:
        conversation_id:
          type: string
          format: uuid
          title: Conversation Id
        messages:
          items:
            $ref: '#/components/schemas/MessageResponse'
          type: array
          title: Messages
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - conversation_id
      - messages
      - total
      - limit
      - offset
      title: ListMessagesResponse
      description: Response for listing messages
...
    SendMessageResponse:
      properties:
        user_message:
          $ref: '#/components/schemas/MessageResponse'
        assistant_message:
          $ref: '#/components/schemas/MessageResponse'
      type: object
      required:
      - user_message
      - assistant_message
      title: SendMessageResponse
      description: Response when sending a message
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token (without 'Bearer' prefix)